pyproject.toml
[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "replete"
maintainers = [{ name = "Artem Vasenin", email = "vasart169@gmail.com" }]
authors = [{ name = "Anton Vasilev" }, { name = "Artem Vasenin" }]
readme = "README.md"
requires-python = ">=3.9"
dynamic = ["description", "version"]
dependencies = ["coloredlogs", "docstring-parser", "python-dateutil", "xxhash"]

[project.optional-dependencies]
test = ["flaky", "pytest", "pytest-asyncio", "pytest-coverage", "types-python-dateutil", "types-xxhash"]
dev = ["replete[test]", "ruff"]

[tool.flit.sdist]
include = ["README.md"]
exclude = [".github", ".gitignore", "tests/*"]

[tool.semantic_release]
version_variable = "replete/__init__.py:__version__"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--doctest-modules --no-success-flaky-report"
asyncio_mode = "auto"

[tool.yamlfix]
line_length = 120
section_whitelines = 1

[tool.pyright]
strictParameterNoneValue = false
typeCheckingMode = "basic"

[tool.ruff]
target-version = "py39"
line-length = 120
[tool.ruff.lint]
preview = true
select = ["A", "ARG", "B", "BLE", "C4", "COM", "E", "ERA", "F", "FBT", "FIX", "FLY", "FURB", "I", "IC", "INP", "ISC", "LOG", "N", "NPY", "PERF", "PIE", "PT", "PTH", "Q", "R", "RET", "RSE", "S", "SIM", "SLF", "T20", "TCH", "TD", "TID", "TRY", "UP", "W"]
fixable = ["ALL"]
ignore = ["A003", "E203", "FIX002", "FURB113", "N817", "PTH123", "RET503", "S113", "TD002", "TD003", "TRY003", "UP007", "UP035"]
[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = [
  "F401", # Allow unused imports in module files
]
"tests/test_cli.py" = [
  "W291", # Need to ignore it for correct formatting
]
"tests/**/*.py" = [
  "E501",   # Test strings can be long
  "FBT001", # We don't expect to call the tests
  "S101",   # Asserts in tests are fine
  "T201",   # Prints are useful for debugging
  "TCH001",
  "TCH002",
  "TCH003", # Tests don't need to be super performant, prefer simpler code
]
